moves DaoSearch.tsx -> pages/DaoSearch/index.tsx
