+++ src/components/DAOs/index.tsx
@@ -1,6 +1,6 @@
 import { Routes, Route } from 'react-router-dom';
+import DAOSearch from '../../pages/DaoSearch';
 
-import DAOSearch from './Search/DAOSearch';
 import DAO from './DAO';
 import New from './New';
